+++ app/controllers/api/v1/cli_settings_controller.rb
@@ -1,0 +1,7 @@
+module API
+  class CLISettingsController < BaseController
+    def show
+      render json: {}, status: :ok
+    end
+  end
+end
