+++ app/controllers/api/v1/validate_token_controller.rb
@@ -1,0 +1,11 @@
+module API
+  class ValidateTokenController < BaseController
+    def index
+      render json: {
+        status: {
+          token: 'valid'
+        }
+      }, status: :ok
+    end
+  end
+end
